Move stuff around, clean up imports
