+++ assets/src/design-system/components/input/test/useNumericInput.js
@@ -20,7 +20,7 @@ import { act, renderHook } from '@testing-library/react-hooks';
 /**
  * Internal dependencies
  */
-import { useNumericInput } from '../numericInput';
+import { useNumericInput } from '../useNumericInput';
 
 const mockOnChange = jest.fn();
